+++ example/example/lib/stories.dart
@@ -1,4 +1,6 @@
-import 'package:example/stories/zwap_buttons.dart';
+import 'package:example/stories/zwap_buttons_story.dart';
+import 'package:example/stories/zwap_checkbox_story.dart';
+import 'package:example/stories/zwap_input_story.dart';
 import 'package:example/stories/zwap_percent.dart';
 import 'package:example/stories/zwap_weekly_calendar_pickert.dart';
 import 'package:flutter/material.dart';
@@ -9,6 +11,8 @@ enum ZwapStories {
   buttons,
   percent,
   weeklyCalendarPicker,
+  checkbox,
+  input,
 }
 
 class ZwapStory {
@@ -38,5 +42,15 @@ List<ZwapStory> stories = [
     name: "ZwapWeeklyCalendarPicker",
     builder: (context) => const ZwapCalendarPickerStory(),
     story: ZwapStories.weeklyCalendarPicker,
+  ),
+  ZwapStory(
+    name: "ZwapCheckBox",
+    builder: (context) => const ZwapCheckBoxStory(),
+    story: ZwapStories.checkbox,
+  ),
+  ZwapStory(
+    name: "ZwapInput",
+    builder: (context) => const ZwapInputStory(),
+    story: ZwapStories.input,
   )
 ];