+++ CHANGELOG.md
@@ -419,3 +419,7 @@
     * ZwapPercentWidget
     * ZwapWeeklyCalendarPicker
 * Removed all `print()´ calls inside components
+
+## 0.1.511
+* Refactoring of ZwapCheckBox
+* Added ZwapCheckBoxDecorations and ZwapCheckBoxStateDecorations and design sysem defaults